requirements.txt
alembic==1.7.7
Authlib==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
ecdsa==0.17.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.20.0
python-jose==3.3.0
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
